+++ Network/Telegram/API/Bot.hs
@@ -3,5 +3,6 @@ module Network.Telegram.API.Bot (module Exports) where
 import Network.Telegram.API.Bot.Property as Exports
 import Network.Telegram.API.Bot.Capacity as Exports
 import Network.Telegram.API.Bot.Object as Exports
+import Network.Telegram.API.Bot.Endpoint as Exports
 import Network.Telegram.API.Bot.Access as Exports
 import Network.Telegram.API.Bot.Core as Exports