+++ Network/Telegram/API/Bot/Object/Chat.hs
@@ -2,11 +2,10 @@ module Network.Telegram.API.Bot.Object.Chat (Chat (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
 import "base" Data.Int (Int64)
-import "base" Control.Applicative (Applicative (pure, (<*>)), Alternative ((<|>)))
+import "base" Control.Applicative (Applicative ((<*>)))
 import "base" Control.Monad (Monad ((>>=)), fail)
-import "base" Data.Function ((.), ($))
-import "base" Data.Functor (Functor (fmap), (<$>))
-import "base" Data.Maybe (Maybe, maybe)
+import "base" Data.Function (($))
+import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
 import "text" Data.Text (Text)