+++ Network/Telegram/API/Bot/Object/Member.hs
@@ -1,13 +1,11 @@
 module Network.Telegram.API.Bot.Object.Member (Member (..)) where
 
 import "aeson" Data.Aeson (FromJSON (parseJSON), withObject, (.:))
-import "base" Control.Applicative (Applicative (pure, (<*>)), Alternative ((<|>)))
+import "base" Control.Applicative (Applicative ((<*>)))
 import "base" Control.Monad (Monad ((>>=)), fail)
-import "base" Data.Function ((.), ($))
-import "base" Data.Functor (Functor (fmap), (<$>))
-import "base" Data.Maybe (Maybe, maybe)
+import "base" Data.Function (($))
+import "base" Data.Functor ((<$>))
 import "base" Text.Show (Show)
-import "text" Data.Text (Text)
 import "text" Data.Text (Text)
 import "time" Data.Time.Clock.POSIX (POSIXTime)